% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_data_item_explorer.R
\name{ng_data_item_explorer_dictionary}
\alias{ng_data_item_explorer_dictionary}
\title{Get the data dictionary for the data item explorer.}
\usage{
ng_data_item_explorer_dictionary()
}
\description{
Get the data dictionary for the data item explorer.
}
\examples{

ng_data_item_explorer_dictionary()


}
