% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_gas_transmission_flows.R
\name{ng_gas_transmission_flows_agg_supply_demand}
\alias{ng_gas_transmission_flows_agg_supply_demand}
\title{Nationalgrid gas download}
\usage{
ng_gas_transmission_flows_agg_supply_demand(date_from = NA, date_to = NA)
}
\description{
Nationalgrid gas download
}
