% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_gas_transmission_flows.R
\name{ng_gas_transmission_hourly_avg}
\alias{ng_gas_transmission_hourly_avg}
\title{Average data to hourly}
\usage{
ng_gas_transmission_hourly_avg(ng_df)
}
\arguments{
\item{ng_df}{data.frame with ng_gas_transmission flow data.}
}
\description{
Average data to hourly
}
