% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_data_item_explorer.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dataItemExplorer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ukgasapi}{\code{\link[ukgasapi]{dataItemExplorer}}}
}}

