% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_gas_transmission_flows.R
\name{ng_gas_transmission_flows_all_zones}
\alias{ng_gas_transmission_flows_all_zones}
\title{Nationalgrid gas download}
\usage{
ng_gas_transmission_flows_all_zones(
  query_type = c("LastUpdate", "Last24Hours", "CustomDates"),
  date_from = NA,
  date_to = NA
)
}
\arguments{
\item{query_type}{The query to be used. Defaults to 'LastUpdate'}

\item{date_from}{Date. In combination with query_type='CustomDates'.}

\item{date_to}{Date. In combination with query_type='CustomDates'.}
}
\description{
Nationalgrid gas download
}
\examples{

library(nationalgridr)

ng_gas_transmission_flows_all_zones(query_type = "LastUpdate")
ng_gas_transmission_flows_all_zones(query_type = "Last24Hours")
ng_gas_transmission_flows_all_zones(query_type = "CustomDates", date_from = Sys.Date()-1, date_to = Sys.Date())

}
